kelp_ma_automation/requirements.txt
python-pptx>=0.6.21
ollama>=0.4.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
pandas>=2.0.0
pyyaml>=6.0
python-docx>=0.8.11
lxml>=4.9.0
requests>=2.31.0
